lambda_bayesian/requirements.txt
boto3==1.28.85
psycopg2-binary==2.9.9
pgmpy==0.1.26
numpy==1.24.3
